remove `UIKit framework dependency` and fix first time data not loading issue
