+++ RandomGeekJoke/RandomGeekJoke/Repository/Repository.swift
@@ -9,7 +9,7 @@ import Foundation
 import CoreData
 
 protocol Repository {
-    func fetch(completion: @escaping LocalStore.RetrievalCompletion) async
+    func fetch(completion: @escaping LocalStore.RetrievalCompletion)
 }
 
 final class JokeRepository: Repository {
@@ -24,18 +24,25 @@ final class JokeRepository: Repository {
         self.remoteFetcher = remoteFetcher
     }
     
-    func fetch(completion: @escaping LocalStore.RetrievalCompletion) async {
-        
-        let result = await fetch()
-        
-        switch result {
-        case .success(let joke):
-            
-            self.removeAndStore(joke: joke)
-            return self.retrieve(completion: completion)
-            
-        case .failure(let error):
-            return self.retrieve(completion: completion)
+    func fetch(completion: @escaping LocalStore.RetrievalCompletion) {
+        Task {
+            let result = await fetch()
+                        
+            switch result {
+            case .success(let joke):
+                
+                self.removeAndStore(joke: joke, completion: { [weak self] error in
+                    if let error = error {
+                        completion(.failure(error))
+                    }
+                    self?.retrieve(completion: completion)
+                })
+                
+            case .failure:
+                //  API Request fail, still looking for data in coredata to retrive previously stored jokes
+                // In real app we can gracefully handle this error by showing Toast meaasage on UI
+                self.retrieve(completion: completion)
+            }
         }
     }
     
@@ -48,32 +55,33 @@ final class JokeRepository: Repository {
         }
     }
     
-    private func removeAndStore(joke: String) {
+    private func removeAndStore(joke: String, completion: @escaping (Error?) -> Void) {
         
         self.retrieve { [weak self] result in
             guard let self = self else { return }
 
             guard case let .found(joke: data) = result, data.count >= JokeRepository.maxStorageAllowed else {
                 
-                self.store(joke: joke)
+                self.store(joke: joke, completion: completion)
                 return
             }
             
             if let first = data.first {
                 self.remove(joke: first, completion: { [weak self] error in
-                    self?.store(joke: joke)
+                    self?.store(joke: joke, completion: completion)
                 })
             }
         }
     }
     
-    @discardableResult
-    private func store(joke: String) -> Error? {
-        var error: Error?
-        self.localStore.insert(joke) { err in
-            error = err
+    private func store(joke: String, completion: @escaping (Error?) -> Void) {
+        self.localStore.insert(joke) { error in
+            if let error = error {
+                completion(error)
+            } else {
+                completion(nil)
+            }
         }
-        return error
     }
     
     private func retrieve(completion: @escaping LocalStore.RetrievalCompletion) {
